+++ tool/src/main/kotlin/com/lightrss/reader/RssScreens.kt
@@ -853,7 +853,7 @@ class SettingsScreen(
                             modifier = Modifier.padding(top = 0.25f.gridUnitsAsDp()),
                         )
                         LightText(
-                            text = "VERSION 1.6.0",
+                            text = "VERSION 1.6.1",
                             variant = LightTextVariant.Superfine,
                             lighten = true,
                             modifier = Modifier.padding(top = 1f.gridUnitsAsDp()),